requirements.txt
absl-py==0.9.0
appdirs==1.4.4
appnope==0.1.0
arrow==0.15.7
attrs==19.3.0
autograd==1.3
backcall==0.2.0
binaryornot==0.4.4
bleach==3.1.5
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
cookiecutter==1.7.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
future==0.18.2
gin-config==0.3.0
gym==0.17.2
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.3.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.1
Jinja2==2.11.2
jinja2-time==0.2.0
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.5
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.2
Mesa==0.8.7
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
nose==1.3.7
notebook==6.0.3
numpy==1.16.5
packaging==20.4
pandas==0.24.2
pandocfilters==1.4.2
parso==0.7.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
ply==3.11
poyo==0.5.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
py-mini-racer==0.3.0
pyglet==1.5.0
Pygments==2.6.1
Pyomo==5.7
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.1
PyUtilib==6.0.0
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
requests==2.24.0
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.10.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
statsmodels==0.11.1
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
torch==1.5.1
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
whynot==0.12.0
widgetsnbextension==3.5.1
zipp==3.1.0
